Only check for child items if temp dir still exists

+++ lib/decompress.ps1
@@ -123,9 +123,12 @@ function Expand-7zipArchive {
     }
     if (!$IsTar -and $ExtractDir) {
         movedir "$DestinationPath\$ExtractDir" $DestinationPath | Out-Null
-        # Remove temporary directory if empty
+        # Remove temporary directory if it still exists and is empty
         $ExtractDirFullPath = [string] "$DestinationPath\$($ExtractDir -replace '[\\/].*')"
-        if ((Get-ChildItem -Path $ExtractDirFullPath -Force).Count -eq 0) {
+        if (
+            (Test-Path -Path $ExtractDirFullPath -PathType 'Container') -and
+            (Get-ChildItem -Path $ExtractDirFullPath -Force).Count -eq 0
+        ) {
             Remove-Item -Path $ExtractDirFullPath -Recurse -Force -ErrorAction Ignore
         }
     }
